Move all File child classes under file.py module
